Tetris blocks cannot any more get stuck when performing a side move
separating handling of side moves from down moves in collision detection
helped to reduce cases we have to check for to properly label the state
of a game field. there was no check wheter a field below a potential
collision is free or not when performing a side walk for all the sub
cases when the next move would result in a collision. the detector
always reported is_blocked and thus the tetris block was placed in the
grid => resulted in placing non-downwards-occluded blocks when hiting sidewards
another blocks.