Fixed approach of determining when a game field is drawable
Previousely, a game field (i.e. a grid cell) was only marked as
"drawable" if was neither colored in black nor in white. this restricted
the color plate that could be used for drawing by those two colors.

Therefore I introduced a new game_field type, called 'moving'. A game
field is moving if it is not placed but should be drawn onto a canvas.